entities/vulnerability:feature - adding new vulnerability fields
Added CWEs, CVEs, Mitigation, Reference, SafeExample, UnsafeExample,
OldHashes into the vulnerability struct. This fields are going to be
used to inform a more depper output from Horusec. OldHashes replaces the
VulnHashInvalid fields, now it's an array that contains all hashes that
are no more generated, but still valid. This is necessary to avoid
breaking changes, but it will be removed in a near future.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>